requirements.txt
vaderSentiment==3.3.2
openai>=1.30.0
python-dotenv>=1.0.0
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
